+++ yulu/scripts/yulu_ui/web/src/App.tsx
@@ -4,6 +4,7 @@ import { useState } from "react";
 import { trpc, makeTrpcClient } from "./trpc.js";
 import { ThemeConfigSync, ThemeProvider } from "./theme.js";
 import { LanguageConfigSync, LanguageProvider } from "./i18n/LanguageProvider.js";
+import { UndoToastProvider, useUndoToast } from "./components/UndoToast.js";
 import { WsProvider } from "./ws.js";
 import { RootLayout } from "./routes/root.js";
 import { InboxLayout, handle as inboxLayoutHandle } from "./routes/inbox/_layout.js";
@@ -108,13 +109,20 @@ export function App() {
         <ThemeProvider>
           <ThemeConfigSync />
           <LanguageProvider>
-            <LanguageConfigSync />
-            <WsProvider>
-              <RouterProvider router={router} />
-            </WsProvider>
+            <UndoToastProvider>
+              <LanguageConfigSyncWithToast />
+              <WsProvider>
+                <RouterProvider router={router} />
+              </WsProvider>
+            </UndoToastProvider>
           </LanguageProvider>
         </ThemeProvider>
       </QueryClientProvider>
     </trpc.Provider>
   );
+}
+
+function LanguageConfigSyncWithToast() {
+  const { showError } = useUndoToast();
+  return <LanguageConfigSync onError={showError} />;
 }